% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/functions.R
\name{best.match}
\alias{best.match}
\title{best.match}
\usage{
best.match(actual, nominal)
}
\arguments{
\item{actual}{value or vector. the values to be matched}

\item{nominal}{value or vector. the values to be matched against}
}
\value{
vector. matched values. values are matched by their minimal absolut distance to its target.
}
\description{
matches values to ones from a given nominal vector by their closest distance to it.
}
\examples{
plot( best.match( dnorm( .05 * -100:100 ), seq( 0, .5, by = .1 ) ) )
}
