% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/functions.R
\name{calc.distr}
\alias{calc.distr}
\title{calc.distr}
\usage{
calc.distr(q, prediction, family = NULL)
}
\arguments{
\item{q}{quantiles}

\item{prediction}{a gamlss prediction}

\item{family}{the family of prediction, for the case it's not an attribute}
}
\value{
z-scores for
}
\description{
not for direct use
}
