% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/functions.R
\name{calc.prob}
\alias{calc.prob}
\title{calc.prob}
\usage{
calc.prob(x, prediction, family = NULL)
}
\arguments{
\item{family}{the family of prediction, for the case it's not an attribute}
}
\description{
not for direct use
}
