% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/functions.R
\name{calc.quant}
\alias{calc.quant}
\title{calc.quant}
\usage{
calc.quant(cent, prediction, family = NULL)
}
\arguments{
\item{cent}{value. centiles that should be calculated. 0.0 < cent & cent < 1.0}

\item{prediction}{a data.frame conraining distribution parameters (mu,sigma,nu,tau) for a certain gamlss model.}

\item{family}{the family of prediction, for the case it's not an attribute}
}
\value{
values relating to probabilities and prediction or family
}
\description{
wrapper: is the quantile function for the respective family or prediction
}
\examples{
plot(d<-data.frame(x=1:1000,y=rBCTo(1000,1000,100,10,1)+rnorm( 1000,0,.1*1:1000)+1:1000))
(m<-lms(y,x,data=d)) # somtimes BCPEo or BCTo
(f<-predictAll(m,data.frame(x=10*c(1:100))))
(p<-as.data.frame(sapply(cnt<-c(.025,.5,.975),calc.quant,f)))
}
