% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/functions.R
\name{compute.percentiles}
\alias{compute.percentiles}
\title{compute.percentiles}
\usage{
compute.percentiles(cent = c(0.025, 0.1, 0.5, 0.9, 0.975), prediction,
  family = NULL)
}
\arguments{
\item{cent}{the centiles}

\item{prediction}{a prediction}

\item{family}{the family of prediction, for the case it's not an attribute}
}
\description{
computes percentiles for a gamlss prediction
}
