% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/functions.R
\name{compute.prediction}
\alias{compute.prediction}
\title{compute.prediction}
\usage{
compute.prediction(x, model)
}
\arguments{
\item{x}{the independent value}

\item{model}{a gamlss model}
}
\value{
a gamlss prediction of the model with new x values
}
\description{
computes a prediction for a gamlss model and new x values
}
