% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/functions.R
\name{compute.references}
\alias{compute.references}
\title{compute.references}
\usage{
compute.references(cent, prediction, family = NULL)
}
\arguments{
\item{prediction}{a gamlss prediction}

\item{family}{the family of prediction, for the case it's not an attribute}
}
\description{
computes reference values for a gamlss prediction
}
