% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/functions.R
\name{compute.sds}
\alias{compute.sds}
\title{compute.sds}
\usage{
compute.sds(model, y = model$y, x = model$xvar)
}
\arguments{
\item{model}{a gamlss model}

\item{y}{the dependent variable}

\item{x}{the independent variable}
}
\value{
the residuals of the model
}
\description{
computes sds for a gamlss model
}
