% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/functions.R
\name{density}
\alias{density}
\title{density}
\usage{
density(x, prediction, family = NULL)
}
\arguments{
\item{x}{independent values}

\item{prediction}{a gamlss prediction}

\item{family}{the family of prediction, for the case it's not an attribute}
}
\value{
z-scores for
}
\description{
not for direct use
}
