% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/functions.R
\name{do.the.whole.thing}
\alias{do.the.whole.thing}
\title{do.the.whole.thing}
\usage{
do.the.whole.thing(cent, y.col.name = col.name, x.col.name = "AGE",
  group.col.name = "SEX", data, x.pred = NULL, fam = c("BCPE", "BCT",
  "BCCG", "BCPEo", "BCTo", "BCCGo"))
}
\arguments{
\item{x.pred}{}
}
\description{
does what it does
}
