% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/functions.R
\name{pDist}
\alias{pDist}
\title{probability}
\usage{
pDist(x, prediction, family = NULL)
}
\arguments{
\item{x}{independent values}

\item{prediction}{a gamlss prediction}

\item{family}{the family of prediction, for the case it's not an attribute}
}
\value{
values relating to probabilities and prediction or family
}
\description{
not for direct use
}
