% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/functions.R
\name{which.best.match}
\alias{which.best.match}
\title{which.best.match}
\usage{
which.best.match(actual, nominal)
}
\arguments{
\item{actual}{value or vector. the values to be matched}

\item{nominal}{value or vector. the values to be matched against}
}
\value{
vector of indices. indeices of matched values in nominal. values are matched by their minimal absolut distance to its target.
}
\description{
matches values to ones from a given nominal vector by their closest distance to it.
}
\examples{
plot( which.best.match( dnorm( .05 * -100:100 ), seq( 0, 1, by = .1 ) ) )
}
