% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/functions.R
\name{compute.model}
\alias{compute.model}
\title{compute.model}
\usage{
compute.model(y, x, families = c("BCCG", "BCPE", "BCT", "BCCGo", "BCPEo",
  "BCTo"), n.cyc = 30, refit = F)
}
\arguments{
\item{y}{y}

\item{x}{x}

\item{families}{families try to fit}

\item{n.cyc}{number of iterations}

\item{refit}{continue computation}
}
\value{
a gamlss model
}
\description{
computes the gamlss model
}
