% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/functions.R
\name{compute.prediction}
\alias{compute.prediction}
\title{compute.prediction}
\usage{
compute.prediction(x, model)
}
\arguments{
\item{model}{}
}
\description{
computes a prediction for a gamlss model and new x values
}
