% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/functions.R
\name{compute.references}
\alias{compute.references}
\title{compute.references}
\usage{
compute.references(cent = c(0.025, 0.1, 0.5, 0.9, 0.975), prediction,
  family = NULL)
}
\arguments{
\item{cent}{the centiles}

\item{prediction}{a prediction}

\item{family}{the family of prediction, for the case it's not an attribute}
}
\value{
a data frame where every centile is a column along x
}
\description{
computes reference values for some percentiles for a gamlss prediction
}
